+++ packages/firecms_core/src/util/icon_list.ts
@@ -1,17 +1,21 @@
 import { iconSynonyms } from "./icon_synonyms";
 import { iconKeys } from "@firecms/ui";
+import Fuse from "fuse.js";
 
-// @ts-ignore
-import * as JsSearch from "js-search";
-
-export const iconsSearch = new JsSearch.Search("key");
-iconsSearch.addIndex("synonyms");
-
-iconsSearch.addDocuments(iconKeys
+const map = iconKeys
     .map((importName) => {
+        // @ts-ignore
+        const iconSynonym = importName in iconSynonyms ? iconSynonyms[importName] : "";
         return {
             key: importName,
-            // @ts-ignore
-            synonyms: iconSynonyms[importName] ?? [],
+            synonyms: iconSynonym,
         }
-    }));
+    });
+export const iconsSearch = new Fuse(map, {
+    isCaseSensitive: false,
+    shouldSort: true,
+    distance: 0,
+    keys: ["key", "synonyms"]
+})
+
+